Leetcode/src/leetcode/no144_BinaryTreePreorderTraversal/SolutionTest.java
package leetcode.no144_BinaryTreePreorderTraversal;

import static org.junit.Assert.*;

import java.util.Arrays;
import java.util.List;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import leetcode.TreeNode;

public class SolutionTest {

	private Solution solution;

	@Before
	public void setUp() throws Exception {
		solution = new Solution();
	}

	@After
	public void tearDown() throws Exception {
		solution = null;
	}

	@Test
	public void initiated() {
		assertNotNull(solution);
	}

	@Test
	public void test() {
		TreeNode root = new TreeNode(1);
		root.right = new TreeNode(2);
		root.right.left = new TreeNode(3);
		List<Integer> expected = Arrays.asList(new Integer[] { 1, 2, 3 });
		List<Integer> actual = solution.preorderTraversal(root);
		assertEquals(expected, actual);
	}

}
